Changed `find_packages()` for `find_namespace_packages()` since translation dir not included as package.
```bash
\############################
\# Package would be ignored #
\############################

Python recognizes 'scipplan.translation' as an importable package,
but it is not listed in the `packages` configuration of setuptools.

'scipplan.translation' has been automatically added to the distribution only
because it may contain data files, but this behavior is likely to change
in future versions of setuptools (and therefore is considered deprecated).

Please make sure that 'scipplan.translation' is included as a package by using
the `packages` configuration field or the proper discovery methods
(for example by using `find_namespace_packages(...)`/`find_namespace:`
instead of `find_packages(...)`/`find:`).

You can read more about "package discovery" and "data files" on setuptools
documentation page.
```